requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
pydantic==2.4.2
python-dotenv==1.0.0
prometheus-client==0.17.1
python-json-logger==2.0.7
requests==2.31.0
eventlet==0.33.3 